Refactor SearchSideBar to take panels as arguments.
