+++ web/src/components/SearchSideBar/SearchSideBar.tsx
@@ -1,53 +1,37 @@
 import React, { useState }  from 'react';
 
 import IconButton           from '../IconButton';
-import panels, { PanelMap } from './panels';
-import TextInput            from '../TextInput';
 import styles               from './SearchSideBar.module.css';
 
 
 interface Props {
-    tags?:        [string, number][];
     initialPanel: string;
-    children?:    PanelMap;
-    editable?:    boolean;
+    children?: {
+        icon: string;
+        name: string;
+        elem: JSX.Element
+    }[];
 }
 
 const SearchSideBar = (props: Props) => {
-    const [search, changeSearch] = useState('');
-    const [panel, changePanel]   = useState<string>(props.initialPanel);
-    const panelMap: PanelMap     = {
-        ...panels,
-        ...props.children
-    };
-
-    const Current               = panelMap[panel].panel;
+    const [panel, changePanel] = useState<string>(props.initialPanel);
+    const current              = props.children && props.children.find(child => child.name === panel)
+    const pannelButtons        = props.children && props.children.map(child =>
+        <IconButton
+            key={child.name}
+            icon={child.icon}
+            active={panel === child.name}
+            onClick={() => changePanel(child.name)}
+        />
+    )
 
     return (
         <div className={styles.Root}>
-            <TextInput
-                value={search}
-                placeholder="Search"
-                onChange={(e) => changeSearch(e.target.value)}
-                onKeyPress={(e) => {
-                }}
-            />
-
             <div className={styles.Buttons}>
-                {
-                    Object.values(panelMap).map(v =>
-                        <IconButton
-                            key={v.name}
-                            icon={v.icon}
-                            tooltip={v.tooltip}
-                            active={panel === v.name}
-                            onClick={() => changePanel(v.name)}
-                        />
-                    )
-                }
+                { pannelButtons }
             </div>
 
-            <Current {...props} />
+            { current && current.elem }
         </div>
     );
 }
