--- web/src/components/SearchSideBar/panels/SettingsPanel.tsx
@@ -1,17 +1,0 @@
-import React  from 'react';
-import Toggle from '../../Toggle';
-
-
-export default {
-    name:    'settings',
-    icon:    'gears',
-    tooltip: 'Settings',
-    panel:   () =>
-        <div>
-            <h1>Settings</h1>
-
-            <Toggle on text="Hide NSFW" onPress={() => {}} />
-            <Toggle on text="Hide Favorites" onPress={() => {}} />
-            <Toggle on text="Only Untagged" onPress={() => {}} />
-        </div>
-};
