+++ web/src/panels/MetaDataPanel/index.tsx
@@ -1,0 +1,2 @@
+import MetaDataPanel from './MetaDataPanel';
+export default MetaDataPanel;
