+++ web/src/panels/SettingsPanel/SettingsPanel.tsx
@@ -1,0 +1,14 @@
+import React  from 'react';
+import Toggle from '../../components/Toggle';
+
+
+const SettingsPanel = () =>
+    <div>
+        <h1>Settings</h1>
+
+        <Toggle on text="Hide NSFW" onPress={() => {}} />
+        <Toggle on text="Hide Favorites" onPress={() => {}} />
+        <Toggle on text="Only Untagged" onPress={() => {}} />
+    </div>;
+
+export default SettingsPanel;
