+++ web/src/panels/SettingsPanel/index.tsx
@@ -1,0 +1,3 @@
+import SettingsPanel from './SettingsPanel';
+export default SettingsPanel;
+
