+++ web/src/panels/TagPanel/index.tsx
@@ -1,0 +1,2 @@
+import TagPanel from './TagPanel';
+export default TagPanel;
